Fixed a bug where GetDisplayName was returning the both the std and dst name separated with a slash if it was '/' separated in the zone FOTMAT column in the TZDB files.
FROM IANA TZDB:

The FORMAT column specifies the usual abbreviation of the time zone name. It can have one of three forms:

a string of three or more characters that are either ASCII alphanumerics, “+”, or “-”, in which case that’s the abbreviation
a pair of strings separated by a slash (‘/’), in which case the first string is the abbreviation for the standard time name and the second string is the abbreviation for the daylight saving time name
a string containing “%s,” in which case the “%s” will be replaced by the text in the appropriate Rule’s LETTER column

Also updated and fixed remaining test case bugs.